# docs/03-api/openapi.yaml
openapi: 3.0.3
info:
  title: Velum indexer API
  version: "0.1.0"
  description: |
    Read-only REST API over the public state of the Velum contract on Midnight.

    This file is written by hand from the route code in `indexer/src/api/` and is the reference
    for the API's shape. It is not generated and is not served by the API. When a route changes,
    change this file in the same commit.

    Data is eventually consistent: it reflects the chain with a delay.
    All identifiers and hashes are 64-character lowercase hex strings without a `0x` prefix.
servers:
  - url: https://velum-api.adasouls.io
    description: Preprod
  - url: http://localhost:3001
    description: Local
...
paths:
...
  /api/events:
    get:
      tags: [events]
      summary: List events
      description: Returns every event, active ones first, then by creation block ascending. Not paginated.
      parameters:
        - name: issuerPk
          in: query
          required: false
          description: Return only events organized by this public key.
          schema:
            $ref: "#/components/schemas/Hex32"
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/events/{eventId}:
    get:
      tags: [events]
      summary: Get one event
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: The event, with its count of non-burned tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventWithLiveTokens"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/events/{eventId}/tokens:
    get:
      tags: [events]
      summary: List an event's tokens
      description: Every token minted for the event, by token id ascending. Not paginated.
      parameters:
        - $ref: "#/components/parameters/EventId"
        - name: includeBurned
          in: query
          required: false
          description: Pass `false` to leave out burned tokens. Any other value, or omitting it, includes them.
          schema:
            type: string
            enum: ["true", "false"]
            default: "true"
      responses:
        "200":
          description: Tokens (an empty array if the event has none)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Token"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Hex32"
  responses:
    NotFound:
      description: No such resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected error. Details are in the server log only.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: internal server error
  schemas:
    Hex32:
      type: string
      pattern: "^[0-9a-f]{64}$"
      description: 32 bytes as 64 lowercase hex characters, no `0x` prefix.
      example: 0aa2e6eaecb0d497f4894627e66bb98c48a579719389c44fd997086168aee0b4
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Event:
      type: object
      required:
        - eventId
        - issuerPk
        - maxSupply
        - expiration
        - isActive
        - isPublicMint
        - metadataURI
        - minted
        - privateAttributesRoot
        - createdBlock
        - createdTx
        - deactivatedBlock
      properties:
        eventId:
          $ref: "#/components/schemas/Hex32"
        issuerPk:
          allOf:
            - $ref: "#/components/schemas/Hex32"
          description: The organizer's public key.
        maxSupply:
          type: integer
          description: Mint cap. 0 means unlimited.
        expiration:
          type: integer
          description: Block time after which minting stops. 0 means never.
        isActive:
          type: boolean
        isPublicMint:
          type: boolean
          description: Whether anyone may claim, or only the organizer may mint.
        metadataURI:
          type: string
          example: ipfs://bafybeih6xhqqfxfyfqgw2xkjxhcxc4kdemoevent/metadata.json
        minted:
          type: integer
          description: Tokens minted so far, burned ones included.
        privateAttributesRoot:
          allOf:
            - $ref: "#/components/schemas/Hex32"
          description: Root of the event's hidden attribute tree. All zeros means none.
        createdBlock:
          type: integer
          nullable: true
        createdTx:
          type: string
          nullable: true
        deactivatedBlock:
          type: integer
          nullable: true
    EventWithLiveTokens:
      allOf:
        - $ref: "#/components/schemas/Event"
        - type: object
          required: [liveTokens]
          properties:
            liveTokens:
              type: integer
              description: Number of non-burned tokens of this event.
    Token:
      type: object
      required:
        - tokenId
        - ownerPk
        - issuerPk
        - firstEventId
        - isBurned
        - mintedBlock
        - mintedTx
        - burnedBlock
        - burnedTx
        - replacesTokenId
        - tokenMetadataURI
        - tokenPrivateMetadataCommit
        - metadataURI
      properties:
        tokenId:
          type: integer
        ownerPk:
          allOf:
            - $ref: "#/components/schemas/Hex32"
          description: The holder's pseudonym for this token's issuer.
        issuerPk:
          $ref: "#/components/schemas/Hex32"
        firstEventId:
          allOf:
            - $ref: "#/components/schemas/Hex32"
          description: The token's event. The name is historical; a token has exactly one event.
        isBurned:
          type: boolean
        mintedBlock:
          type: integer
          nullable: true
        mintedTx:
          type: string
          nullable: true
        burnedBlock:
          type: integer
          nullable: true
        burnedTx:
          type: string
          nullable: true
        replacesTokenId:
          type: integer
          nullable: true
          description: >-
            The token this one replaced, when it was minted by reissueCredential. null otherwise.
        tokenMetadataURI:
          type: string
          nullable: true
          description: This token's own metadata. Prefer it over `metadataURI` when rendering.
        tokenPrivateMetadataCommit:
          type: string
          nullable: true
          description: Commitment to hidden per-token metadata, hex. All zeros means none.
        metadataURI:
          type: string
          nullable: true
          description: The parent event's metadata.
...